refactor: moved source code under lib and global headers under include
