+++ firmware/platformio.ini
@@ -22,5 +22,6 @@ lib_deps=
 
 board_build.mcu = atmega32u4
 board_build.f_cpu = 8000000L
+build_flags = -I include
 upload_port=COM12
 monitor_speed=115200
